requirements.txt
fastapi
uvicorn
opencv-python-headless
facenet-pytorch
torch
firebase-admin
numpy
python-multipart
Pillow
jinja2